+++ tests/test_cli.py
@@ -12,6 +12,16 @@ try:
 except ImportError:
     xarray = None
 
+msg = ""
+try:
+    import discretize
+    # Backwards compatibility; remove latest for version 1.0.0.
+    dv = discretize.__version__.split('.')
+    if int(dv[0]) == 0 and int(dv[1]) < 6:
+        msg = "`emg3d>=v0.15.0` ONLY works with `discretize>=v0.6.0`;"
+except ImportError:
+    pass
+
 
 @pytest.mark.script_launch_mode('subprocess')
 def test_basic(script_runner):
@@ -22,7 +32,11 @@ def test_basic(script_runner):
     ret = script_runner.run('emg3d', '-h')
     assert ret.success
     assert "emg3d is a multigrid solver for 3D EM diffusion" in ret.stdout
-    assert ret.stderr == ''
+
+    if msg:
+        assert msg in ret.stderr
+    else:
+        assert ret.stderr == ""
 
     # Test the installed version fails if called without anything.
     ret = script_runner.run('emg3d')
@@ -51,6 +65,9 @@ def test_basic(script_runner):
     assert "CONFIGURATION FILE NOT FOUND" in ret.stderr
 
     os.environ["NUMBA_DISABLE_JIT"] = "0"
+
+    # Clean up
+    os.remove('emg3d_out.log')
 
 
 class TestParser:
@@ -341,7 +358,7 @@ class TestRun:
 
     # Create a dummy grid and model.
     xx = np.ones(16)*500
-    grid = emg3d.TensorMesh([xx, xx, xx], x0=np.array([0, 0, 0]))
+    grid = emg3d.TensorMesh([xx, xx, xx], origin=np.array([0, 0, 0]))
     model = emg3d.Model(grid, 1.)
 
     def test_basic(self, tmpdir, capsys):