+++ tests/test_fields.py
@@ -8,6 +8,10 @@ from numpy.testing import assert_allclose, assert_array_equal
 # Import soft dependencies.
 try:
     import discretize
+    # Backwards compatibility; remove latest for version 1.0.0.
+    dv = discretize.__version__.split('.')
+    if int(dv[0]) == 0 and int(dv[1]) < 6:
+        discretize = None
 except ImportError:
     discretize = None
 
@@ -71,7 +75,7 @@ def test_get_source_field(capsys):
     assert_allclose(sfield.smu0, -iomegamu)
 
     # Put source on final node, should still work.
-    src = [grid.vectorNx[-1], grid.vectorNy[-1], grid.vectorNz[-1],
+    src = [grid.nodes_x[-1], grid.nodes_y[-1], grid.nodes_z[-1],
            src[3], src[4]]
     sfield = fields.get_source_field(grid, src, freq)
     tot_field = np.linalg.norm(
@@ -179,8 +183,8 @@ def test_get_source_field_point_vs_finite(capsys):
     h1 = get_h(4, 2, 200, 1.1)
     h2 = get_h(4, 2, 200, 1.2)
     h3 = get_h(4, 2, 200, 1.2)
-    x0 = np.array([-h1.sum()/2, -h2.sum()/2, -h3.sum()/2])
-    grid3 = meshes.TensorMesh([h1, h2, h3], x0)
+    origin = np.array([-h1.sum()/2, -h2.sum()/2, -h3.sum()/2])
+    grid3 = meshes.TensorMesh([h1, h2, h3], origin)
     slen = 333
     strength = np.pi
     d_src, f_src = get_f_src([0, 0., 0., 50, 33], slen)
@@ -461,9 +465,13 @@ def test_get_receiver_response():
     src = (0, 0, 0, 0, 0)
     freq = 10
 
-    hx, x0 = meshes.get_hx_h0(freq, res, [0, 1000], min_width=20, verb=0)
-    hyz, yz0 = meshes.get_hx_h0(freq, res, [-25, 25], min_width=20, verb=0)
-    grid = meshes.TensorMesh([hx, hyz, hyz], x0=np.array([x0, yz0, yz0]))
+    grid = meshes.construct_mesh(
+            frequency=freq,
+            center=(0, 0, 0),
+            properties=res,
+            domain=[[0, 1000], [-25, 25], [-25, 25]],
+            min_width_limits=20,
+    )
 
     model = models.Model(grid, res)
     sfield = fields.get_source_field(grid, src, freq)